Changes to Horde Resistance and Player Archetype
This commit contains small code alternations, renaming the "algorithm" function to "algorithm_fun" to make sure that no errors occur with the class and the function, as they shared the same name.

Alterations to the resistance to make sure that as the number of hordes increases, the stronger they become, as well as decreasing the bonus from the level of the player.

Changes to the damage of the player, factoring in the resistance of the hordes.