Algorithm Code
This commit contains new code in the algorithm function in the algorithm class.

It takes into considerations the fight and flight optins. while also removing the previous code in the algorithm.

Additions to he "attack_dic" have been made to include the new categories of stamina and running.

Furthermore, some minor alterations to the code were made.