Intelligence Skill Option Code
This commit contains the code for the player to choose an option of defeating the horde by outwitting them.

Other modificatinos to previous choices, such as adding the intelligence option to if a player fails to outrun the horde have been add.

Intelligence and surrender lists have been added to the dictionary.