Algorithm Alternations
This commit creates multiple functions which contain the code that was previously hard-coded into the algorithm.

This will make the code much more modular than before.

An issue that has to be resolved though, is the modification of the "intelligence" function. As of right now there are two, one if intelligence is the first option, and if intelligence is the second option.

Changing it will make the algorithm shorter and simpler, thus improving if for future use.