Bug Fixing
This commit contains new additions to the code that fix dictionary error.

Furthermore, it also contains new additions to the code (for loop), which fixes the error of not getting the right outcome when the player writes a command.

This was done by writing a new for loop after the one in the algorithm, and changing the content inside of the previous if statements to Boolean Varaibles.