Leveling up System
This commit contains an addition to the Algorithm which a function which can be used to level up the player.

It has yet to be added to algorithm, as a method of determining how much xp a player should get is still being thought out, but the foundations required for the system are already in place.