Horde Drops Code
This commit adds the feature of item drops in the algorithm, as well as how they affect the damage of the player.

To improve the scaling factor of the items, as the player goes on, they'll be able to refine their items if they get the same item as before.

Bugs were fixed as well.